% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jet.R
\name{jet_add}
\alias{jet_add}
\title{Sum of Jets}
\usage{
jet_add(a, b)
}
\arguments{
\item{a, b}{Jets, or a jet and a single number.}
}
\value{
A jet.
}
\description{
Adds two jets, or a jet and a number, componentwise.
}
\keyword{internal}
