% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jet.R
\name{jet_const}
\alias{jet_const}
\title{A Constant Jet}
\usage{
jet_const(v, lay)
}
\arguments{
\item{v}{The value.}

\item{lay}{A layout from \code{\link{jet_layout}}.}
}
\value{
A jet.
}
\description{
A number with every derivative zero.
}
\keyword{internal}
