% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jet.R
\name{jet_digamma}
\alias{jet_digamma}
\title{The Digamma of a Jet}
\usage{
jet_digamma(a, lay)
}
\arguments{
\item{a}{A jet.}

\item{lay}{A layout from \code{\link{jet_layout}}.}
}
\value{
A jet.
}
\description{
\eqn{\psi(a)}, with the polygamma functions above it.
}
\seealso{
\code{\link{jet_compose}}
}
\keyword{internal}
