% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jet.R
\name{jet_div}
\alias{jet_div}
\title{The Quotient of Two Jets}
\usage{
jet_div(a, b, lay)
}
\arguments{
\item{a, b}{Jets over the same layout.}

\item{lay}{A layout from \code{\link{jet_layout}}.}
}
\value{
A jet.
}
\description{
\eqn{a/b}, as a product with the reciprocal.
}
\seealso{
\code{\link{jet_mul}}
}
\keyword{internal}
