% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jet.R
\name{jet_inv}
\alias{jet_inv}
\title{The Reciprocal of a Jet}
\usage{
jet_inv(a, lay)
}
\arguments{
\item{a}{A jet.}

\item{lay}{A layout from \code{\link{jet_layout}}.}
}
\value{
A jet.
}
\description{
\eqn{1/a}, the common case of \code{\link{jet_pow}}.
}
\seealso{
\code{\link{jet_pow}}
}
\keyword{internal}
