% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jet.R
\name{jet_log}
\alias{jet_log}
\title{The Logarithm of a Jet}
\usage{
jet_log(a, lay)
}
\arguments{
\item{a}{A jet.}

\item{lay}{A layout from \code{\link{jet_layout}}.}
}
\value{
A jet.
}
\description{
\eqn{\log a}, with derivatives \eqn{(-1)^{k-1}(k-1)!/a^k}.
}
\seealso{
\code{\link{jet_compose}}
}
\keyword{internal}
