% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jet.R
\name{jet_sqrt}
\alias{jet_sqrt}
\title{The Square Root of a Jet}
\usage{
jet_sqrt(a, lay)
}
\arguments{
\item{a}{A jet.}

\item{lay}{A layout from \code{\link{jet_layout}}.}
}
\value{
A jet.
}
\description{
\eqn{\sqrt{a}}, the common case of \code{\link{jet_pow}}.
}
\seealso{
\code{\link{jet_pow}}
}
\keyword{internal}
