% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jet.R
\name{jet_trigamma}
\alias{jet_trigamma}
\title{The Trigamma of a Jet}
\usage{
jet_trigamma(a, lay = a$lay)
}
\arguments{
\item{a}{A jet.}

\item{lay}{A layout from \code{\link{jet_layout}}; taken from the jet
itself unless given.}
}
\value{
A jet.
}
\description{
\eqn{\psi'(a)}, with the polygamma functions above it.
}
\seealso{
\code{\link{jet_compose}}
}
\keyword{internal}
