% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_fallbacks.R
\name{numerical_dmatrix}
\alias{numerical_dmatrix}
\title{Numerical First Derivatives of a Structure's Matrix}
\usage{
numerical_dmatrix(s, eta)
}
\arguments{
\item{s}{A \code{\link{covstruct}} object.}

\item{eta}{A numeric vector of free values.}
}
\value{
A named list of symmetric matrices.
}
\description{
One central difference of \code{\link{struct_matrix}} in each component of
\eqn{\eta}.
}
\seealso{
\code{\link{struct_dmatrix}}
}
