% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound_symmetry.R
\name{param_value.CompoundSymmetryParam}
\alias{param_value.CompoundSymmetryParam}
\title{Value of a Compound Symmetry Parameter}
\arguments{
\item{s}{A \code{\link[=CompoundSymmetryParam]{CompoundSymmetryParam()}} object.}

\item{eta}{A numeric vector of two free values.}

\item{...}{Unused.}
}
\value{
A compound-symmetric positive definite matrix.
}
\description{
\eqn{\sigma^2\{(1-\rho)I + \rho J\}}.
}
\keyword{internal}
