% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jet.R
\name{print.jet}
\alias{print.jet}
\title{Print a Jet}
\usage{
\method{print}{jet}(x, ...)
}
\arguments{
\item{x}{A jet.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Shows the value and how many variables the derivatives are taken in, rather
than the whole list of components, which is thirty-five numbers even in two
variables.
}
\seealso{
\code{\link{Ops.jet}}
}
\keyword{internal}
