% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_fallbacks.R
\name{struct_d2matrix.covstruct}
\alias{struct_d2matrix.covstruct}
\title{Default Second Derivatives}
\arguments{
\item{s}{A \code{\link{covstruct}} object.}

\item{eta}{A numeric vector of free values.}

\item{...}{Unused.}
}
\value{
A named list of symmetric matrices.
}
\description{
Fallback: see \code{\link{numerical_d2matrix}}.
}
\keyword{internal}
