% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{struct_dmatrix}
\alias{struct_dmatrix}
\title{First Derivatives of a Structure's Matrix}
\usage{
struct_dmatrix(s, eta, ...)
}
\arguments{
\item{s}{An object inheriting from class \code{\link{covstruct}}.}

\item{eta}{A numeric vector of length \code{s@n_free}.}

\item{...}{Passed to methods.}
}
\value{
A list of \code{s@n_free} symmetric matrices, named by
\code{s@free_names}.
}
\description{
Returns \eqn{\partial M / \partial \eta_k} for every free value, as a list
of matrices.
}
\details{
A structure that registers no method for this generic gets the numerical one
of the \code{\link{covstruct}} class, which applies a single central
difference to \code{\link{struct_matrix}} in each component.
}
\examples{
s <- scaled_struct(diag(2))
struct_dmatrix(s, 0)

}
\seealso{
\code{\link{struct_d2matrix}}, \code{\link{struct_is_numerical}}
}
