% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_struct.R
\name{struct_factor.DiagStruct}
\alias{struct_factor.DiagStruct}
\title{Factor of a Diagonal Structure}
\arguments{
\item{s}{A \code{\link{DiagStruct}} object.}

\item{eta}{A numeric vector of free values.}

\item{...}{Unused.}
}
\value{
A diagonal numeric matrix.
}
\description{
The square roots of the entries, on the diagonal.
}
\keyword{internal}
