% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{struct_pair_names}
\alias{struct_pair_names}
\title{Names of the Distinct Second-Derivative Components}
\usage{
struct_pair_names(s)
}
\arguments{
\item{s}{An object inheriting from class \code{\link{covstruct}}.}
}
\value{
A character vector of length \eqn{d(d+1)/2}.
}
\description{
The keys of \code{\link{struct_d2matrix}} and
\code{\link{struct_d2logdet}}: one per unordered pair of free values,
diagonal first and then the off-diagonal pairs in lexicographic order.
}
\details{
This exists so that nothing has to recover a pair by splitting a key apart.
Generating the keys and the index pairs from one enumeration cannot be
fooled by a free name that contains the separator, which splitting can.
Use \code{\link{struct_pair_indices}} for the pairs themselves.
}
\examples{
struct_pair_names(log_cholesky(2))

}
\seealso{
\code{\link{struct_pair_indices}}
}
