% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar1.R
\name{ar1_logdet_terms}
\alias{ar1_logdet_terms}
\title{The Log-Determinant Terms of an AR(1) Parameter}
\usage{
ar1_logdet_terms(s)
}
\arguments{
\item{s}{An \code{\link[=Ar1Param]{Ar1Param()}} object.}
}
\value{
A list of numeric triples.
}
\description{
The affine-logarithm terms of \eqn{q(\rho) = (p-1)\log(1-\rho^2)}, split as
\eqn{(p-1)\{\log(1-\rho) + \log(1+\rho)\}} so that both pieces are
logarithms of affine functions.
}
\keyword{internal}
