% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_log.R
\name{combinat_perms}
\alias{combinat_perms}
\title{All Orderings of a Tuple}
\usage{
combinat_perms(x)
}
\arguments{
\item{x}{An integer vector.}
}
\value{
A list of integer vectors.
}
\description{
The distinct permutations of an index tuple, as a list.
}
\keyword{internal}
