% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound_symmetry.R
\name{cs_logdet_terms}
\alias{cs_logdet_terms}
\title{The Log-Determinant Terms of a Compound-Symmetric Parameter}
\usage{
cs_logdet_terms(s)
}
\arguments{
\item{s}{A \code{\link{CompoundSymmetryParam}} object.}
}
\value{
A list of numeric triples.
}
\description{
The affine-logarithm terms of \eqn{q(\rho) = \log\{1 + (p-1)\rho\} +
(p-1)\log(1-\rho)}, the correlation's half of the log-determinant.
}
\keyword{internal}
