% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jet.R
\name{jet_exp}
\alias{jet_exp}
\title{The Exponential of a Jet}
\usage{
jet_exp(a, lay = a$lay)
}
\arguments{
\item{a}{A jet.}

\item{lay}{A layout from \code{\link{jet_layout}}; taken from the jet
itself unless given.}
}
\value{
A jet.
}
\description{
\eqn{e^a}, whose four derivatives are all \eqn{e^a}.
}
\seealso{
\code{\link{jet_compose}}
}
\keyword{internal}
