% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jet.R
\name{jet_lgamma}
\alias{jet_lgamma}
\title{The Log-Gamma of a Jet}
\usage{
jet_lgamma(a, lay = a$lay)
}
\arguments{
\item{a}{A jet.}

\item{lay}{A layout from \code{\link{jet_layout}}; taken from the jet
itself unless given.}
}
\value{
A jet.
}
\description{
\eqn{\log\Gamma(a)}, whose derivatives are the polygamma functions.
}
\seealso{
\code{\link{jet_compose}}
}
\keyword{internal}
