% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_fallbacks.R
\name{numerical_d1}
\alias{numerical_d1}
\title{Numerical First Derivatives of a Parameter's Matrix}
\usage{
numerical_d1(s, eta)
}
\arguments{
\item{s}{A \code{\link[=parameter]{parameter()}} object.}

\item{eta}{A numeric vector of free values.}
}
\value{
A named list of symmetric matrices.
}
\description{
One central difference of \code{\link[=param_value]{param_value()}} in each component of
\eqn{\eta}.
}
\examples{
numerical_d1(scalar_matrix(2), 0.3)
}
\seealso{
\code{\link[=param_d1]{param_d1()}}
}
