% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{param_d1}
\alias{param_d1}
\title{First Derivatives of a Parameter's Matrix}
\usage{
param_d1(s, eta, ...)
}
\arguments{
\item{s}{An object inheriting from class \code{\link{parameter}}.}

\item{eta}{A numeric vector of length \code{s@n_free}.}

\item{...}{Passed to methods.}
}
\value{
A list of \code{s@n_free} symmetric matrices, named by
\code{s@free_names}.
}
\description{
Returns \eqn{\partial M / \partial \eta_k} for every free value, as a list
of matrices.
}
\details{
A parameter that registers no method for this generic gets the numerical one
of the \code{\link{parameter}} class, which applies a single central
difference to \code{\link{param_value}} in each component.
}
\examples{
s <- scaled_matrix(diag(2))
param_d1(s, 0)

}
\seealso{
\code{\link{param_d2}}, \code{\link{param_is_numerical}}
}
