% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled_matrix.R
\name{param_d1.ScaledMatrixParam}
\alias{param_d1.ScaledMatrixParam}
\title{First Derivative of a Scaled Parameter}
\arguments{
\item{s}{A \code{\link{ScaledMatrixParam}} object.}

\item{eta}{A numeric vector of free values.}

\item{...}{Unused.}
}
\value{
A named list with one symmetric matrix.
}
\description{
Closed form: \eqn{h'(\eta) P}.
}
\keyword{internal}
