% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_fallbacks.R
\name{param_d2.parameter}
\alias{param_d2.parameter}
\title{Default Second Derivatives}
\arguments{
\item{s}{A \code{\link{parameter}} object.}

\item{eta}{A numeric vector of free values.}

\item{...}{Unused.}
}
\value{
A named list of symmetric matrices.
}
\description{
Fallback: see \code{\link{numerical_d2}}.
}
\keyword{internal}
