% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex.R
\name{param_d2.SimplexParam}
\alias{param_d2.SimplexParam}
\title{Second Derivatives of a Simplex Parameter}
\arguments{
\item{s}{A \code{\link[=SimplexParam]{SimplexParam()}} object.}

\item{eta}{A numeric vector of free values.}

\item{...}{Unused.}
}
\value{
A named list of vectors of length \eqn{K}.
}
\description{
Closed form, one further product rule.
}
\keyword{internal}
