% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{param_d2logdet}
\alias{param_d2logdet}
\title{Hessian of the Log-Determinant}
\usage{
param_d2logdet(s, eta, ...)
}
\arguments{
\item{s}{An object inheriting from class \code{\link{parameter}}.}

\item{eta}{A numeric vector of length \code{s@n_free}.}

\item{...}{Passed to methods.}
}
\value{
A named numeric vector, keyed as \code{param_tuple_names(s)}.
}
\description{
Returns the distinct second derivatives of the log-determinant, or of the
log pseudo-determinant, keyed as \code{\link{param_tuple_names}}.
}
\examples{
param_d2logdet(log_cholesky(2), c(0.2, -0.1, 0.4))

}
\seealso{
\code{\link{param_dlogdet}}
}
