% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled_matrix.R
\name{param_d2logdet.ScaledMatrixParam}
\alias{param_d2logdet.ScaledMatrixParam}
\title{Log-Determinant Hessian of a Scaled Parameter}
\arguments{
\item{s}{A \code{\link{ScaledMatrixParam}} object.}

\item{eta}{A numeric vector of free values.}

\item{...}{Unused.}
}
\value{
A named numeric vector.
}
\description{
Closed form: \eqn{r\{h''/h - (h'/h)^2\}}, which under the default log link
is zero, the log pseudo-determinant being linear in the free value.
}
\keyword{internal}
