% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{param_d3logdet}
\alias{param_d3logdet}
\alias{param_d4logdet}
\title{Third and Fourth Derivatives of the Log-Determinant}
\usage{
param_d3logdet(s, eta, ...)

param_d4logdet(s, eta, ...)
}
\arguments{
\item{s}{An object inheriting from class \code{\link{matrix_parameter}}.}

\item{eta}{A numeric vector of length \code{s@n_free}.}

\item{...}{Passed to methods.}
}
\value{
A named numeric vector.
}
\description{
The higher derivatives of the log-(pseudo-)determinant, keyed as
\code{\link{param_tuple_names}} of the matching order.
}
\examples{
param_d3logdet(log_cholesky(2), c(0.1, -0.2, 0.4))

}
\seealso{
\code{\link{param_d2logdet}}
}
