% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_fallbacks.R
\name{param_d4.parameter}
\alias{param_d4.parameter}
\title{Default Fourth Derivatives}
\arguments{
\item{s}{A \code{\link[=parameter]{parameter()}} object.}

\item{eta}{A numeric vector of free values.}

\item{...}{Unused.}
}
\value{
A named list.
}
\description{
Fallback: see \code{\link[=numerical_d4]{numerical_d4()}}.
}
\keyword{internal}
