% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled_matrix.R
\name{param_d4.ScaledMatrixParam}
\alias{param_d4.ScaledMatrixParam}
\title{Fourth Derivatives of a Scaled Parameter}
\arguments{
\item{s}{A \code{\link[=ScaledMatrixParam]{ScaledMatrixParam()}} object.}

\item{eta}{A numeric vector with one free value.}

\item{...}{Unused.}
}
\value{
A named list with one matrix.
}
\description{
Closed form: \eqn{h''''(\eta)\, M_0}.
}
\keyword{internal}
