% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex.R
\name{param_d4.SimplexParam}
\alias{param_d4.SimplexParam}
\title{Fourth Derivatives of a Simplex Parameter}
\arguments{
\item{s}{A \code{\link{SimplexParam}} object.}

\item{eta}{A numeric vector of free values.}

\item{...}{Unused.}
}
\value{
A named list of vectors of length \eqn{K}.
}
\description{
Closed form, the cumulant recursion at fourth order.
}
\keyword{internal}
