% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_matrix.R
\name{param_d4.TransitionMatrixParam}
\alias{param_d4.TransitionMatrixParam}
\title{Fourth Derivatives of a Transition Matrix Parameter}
\arguments{
\item{s}{A \code{\link[=TransitionMatrixParam]{TransitionMatrixParam()}} object.}

\item{eta}{A numeric vector of free values.}

\item{...}{Unused.}
}
\value{
A named list of matrices.
}
\description{
Closed form, row by row.
}
\keyword{internal}
