% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_log.R
\name{param_d4logdet.MatrixLogParam}
\alias{param_d4logdet.MatrixLogParam}
\title{Fourth Log-Determinant Derivatives of a Matrix Logarithm Parameter}
\arguments{
\item{s}{A \code{\link{MatrixLogParam}} object.}

\item{eta}{A numeric vector of free values.}

\item{...}{Unused.}
}
\value{
A named numeric vector of zeros.
}
\description{
Closed form: zero.
}
\keyword{internal}
