% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagonal_matrix.R
\name{param_factor.DiagMatrixParam}
\alias{param_factor.DiagMatrixParam}
\title{Factor of a Diagonal Parameter}
\arguments{
\item{s}{A \code{\link{DiagMatrixParam}} object.}

\item{eta}{A numeric vector of free values.}

\item{...}{Unused.}
}
\value{
A diagonal numeric matrix.
}
\description{
The square roots of the entries, on the diagonal.
}
\keyword{internal}
