% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{param_factor}
\alias{param_factor}
\title{A Factor of a Parameter's Matrix}
\usage{
param_factor(s, eta, ...)
}
\arguments{
\item{s}{An object inheriting from class \code{\link{parameter}}.}

\item{eta}{A numeric vector of length \code{s@n_free}.}

\item{...}{Passed to methods.}
}
\value{
A lower triangular numeric matrix with \code{s@dimension} rows and
columns.
}
\description{
Returns a lower triangular \eqn{L} with \eqn{M = L L^\top}.
}
\details{
Refused for a rank-deficient family, for the reason given in
\code{\link{param_solve}}.
}
\examples{
round(param_factor(log_cholesky(2), c(0.1, 0.2, -0.3)), 4)

}
\seealso{
\code{\link{param_solve}}
}
