% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagonal_matrix.R
\name{param_logdet.DiagMatrixParam}
\alias{param_logdet.DiagMatrixParam}
\title{Log-Determinant of a Diagonal Parameter}
\arguments{
\item{s}{A \code{\link{DiagMatrixParam}} object.}

\item{eta}{A numeric vector of free values.}

\item{...}{Unused.}
}
\value{
A single number.
}
\description{
The sum of the logs of the entries.
}
\keyword{internal}
