% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar1.R
\name{param_value.Ar1Param}
\alias{param_value.Ar1Param}
\title{Value of an AR(1) Parameter}
\arguments{
\item{s}{An \code{\link{Ar1Param}} object.}

\item{eta}{A numeric vector of two free values.}

\item{...}{Unused.}
}
\value{
An AR(1) positive definite matrix.
}
\description{
\eqn{\sigma^2 \rho^{\lvert i-j \rvert}}.
}
\keyword{internal}
