% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoregressive.R
\name{param_value.AutoregressiveParam}
\alias{param_value.AutoregressiveParam}
\title{Value of an Autoregressive Parameter}
\arguments{
\item{s}{An \code{\link{AutoregressiveParam}} object.}

\item{eta}{A numeric vector of free values.}

\item{...}{Unused.}
}
\value{
A positive definite Toeplitz matrix.
}
\description{
The Toeplitz matrix \eqn{\gamma_0 \rho_{\lvert i-j \rvert}}, the
autocorrelations coming from the Levinson-Durbin recursion.
}
\keyword{internal}
