% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_diag.R
\name{param_value.BlockDiagParam}
\alias{param_value.BlockDiagParam}
\title{Value of a Block-Diagonal Parameter}
\arguments{
\item{s}{A \code{\link{BlockDiagParam}} object.}

\item{eta}{A numeric vector of length \code{s@n_free}.}

\item{...}{Unused.}
}
\value{
A symmetric matrix of side \code{s@dimension}.
}
\description{
Assembles the blocks along the diagonal.
}
\keyword{internal}
