% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagonal_matrix.R
\name{param_value.DiagMatrixParam}
\alias{param_value.DiagMatrixParam}
\title{Matrix of a Diagonal Parameter}
\arguments{
\item{s}{A \code{\link{DiagMatrixParam}} object.}

\item{eta}{A numeric vector of free values.}

\item{...}{Unused.}
}
\value{
A diagonal positive definite matrix.
}
\description{
The link applied entrywise, on the diagonal.
}
\keyword{internal}
