% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled_matrix.R
\name{param_value.ScaledMatrixParam}
\alias{param_value.ScaledMatrixParam}
\title{Matrix of a Scaled Parameter}
\arguments{
\item{s}{A \code{\link{ScaledMatrixParam}} object.}

\item{eta}{A numeric vector of free values.}

\item{...}{Unused.}
}
\value{
A symmetric positive semidefinite matrix.
}
\description{
The fixed matrix times the scale.
}
\keyword{internal}
