% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_struct.R
\name{param_value.SumStructParam}
\alias{param_value.SumStructParam}
\title{Value of a Sum of Fixed Matrices}
\arguments{
\item{s}{A \code{\link{SumStructParam}} object.}

\item{eta}{A numeric vector of length \code{s@n_free}.}

\item{...}{Unused.}
}
\value{
A symmetric matrix of side \code{s@dimension}.
}
\description{
The weighted sum of the components.
}
\keyword{internal}
