% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex.R
\name{SimplexParam}
\alias{SimplexParam}
\title{Simplex Parameter}
\usage{
SimplexParam(
  param_name = character(0),
  n_free = integer(0),
  free_names = character(0),
  param_params = list()
)
}
\arguments{
\item{param_name}{A single character string naming the family.}

\item{n_free}{The length \eqn{d} of the free vector.}

\item{free_names}{A character vector of length \code{n_free}.}

\item{param_params}{A list of whatever the family needs to evaluate itself.}
}
\value{
An object of class \code{SimplexParam}.
}
\description{
The S7 class of probability vectors on the open simplex, in the additive
log-ratio parametrisation. Constructed by \code{\link{simplex}}.
}
\examples{
S7::S7_inherits(simplex(3), SimplexParam)

}
\seealso{
\code{\link{simplex}}
}
