% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{tuple_indices}
\alias{tuple_indices}
\title{The Index Tuples of a Given Width}
\usage{
tuple_indices(d, order = 2L)
}
\arguments{
\item{d}{The number of variables.}

\item{order}{The derivative order, 1 to 4.}
}
\value{
A list of integer vectors of length \code{order}.
}
\description{
The enumeration behind \code{\link[=param_tuple_indices]{param_tuple_indices()}}, taken over a
number of variables rather than over a parameter, so that anything holding
derivatives over \eqn{d} variables -- a jet, for instance -- can share it.
}
\seealso{
\code{\link[=param_tuple_indices]{param_tuple_indices()}}
}
\keyword{internal}
